+++ lib/core/constants/sizes.dart
@@ -3,3 +3,4 @@
 const tDefaultSize = 30.0;
 const tSplashContainerSize = 30.0;
 const tButtonHeight = 15.0;
+const tFormHeight = 30.0;
