+++ lib/core/constants/text_strings.dart
@@ -1,25 +1,36 @@
-//App Constant Strings
-
+//Splash Screen Texts
 const String tAppName = "Rutas Turísticas";
 const String tAppTag = "Mancomunidad Bosque Seco";
 
+//Onboarding Screen Texts
 const String tAppOnboardingTitle1 = "Bienvenido";
 const String tAppOnboardingSubTittle1 =
     "A la aplicación de rutas turísticas de la Mancomunidad Bosque Seco";
 const String tAppOnboardingCounter1 = "1/3";
-
 const String tAppOnboardingTitle2 = "Descubre";
 const String tAppOnboardingSubTittle2 =
     "Las funcionalidades de la App y crea tus propias rutas";
 const String tAppOnboardingCounter2 = "2/3";
-
 const String tAppOnboardingTitle3 = "Explora";
 const String tAppOnboardingSubTittle3 = "Las rutas y encuentra nuevos destinos";
 const String tAppOnboardingCounter3 = "3/3";
 
+//Welcome Screen Texts
 const String tAppWelcomeScreenTitle = "RTMB APP";
 const String tWelcomeSubTitle =
     "Mancomunidad de municipalidades del sur occidente de la provincia de Loja “Bosque Seco”";
 
+//Default Texts
 const String tLogin = "Ingresar";
 const String tRegister = "Registrarse";
+
+//Login Screen Texts
+const String tLoginTitle = "Bienvenido de vuelta";
+const String tLoginSubtitle = "Es un placer tenerte de regreso";
+const String tRememberMe = "Recuerdame";
+const String tAlreadyHaveAnAccount = "¿Ya tienes una cuenta?";
+const String tEmail = "E-Mail";
+const String tPassword = "Contraseña";
+const String tForgotPassword = "¿Olvidaste tu contraseña?";
+const String tSingInWithGoogle = "Ingresa con google";
+const String tDontHaveAnAccount = "¿Aún no tienes una cuenta?";
