+++ lib/features/authentication/login/login_form.dart
@@ -1,0 +1,55 @@
+import 'package:flutter/material.dart';
+
+import '../../../core/constants/sizes.dart';
+import '../../../core/constants/text_strings.dart';
+
+class LoginFormWIdget extends StatelessWidget {
+  const LoginFormWIdget({
+    Key? key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Form(
+        child: Container(
+      padding: const EdgeInsets.symmetric(vertical: tFormHeight - 10),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          TextFormField(
+            decoration: const InputDecoration(
+              prefixIcon: Icon(Icons.person_outline_outlined),
+              labelText: tEmail,
+              hintText: tEmail,
+              border: OutlineInputBorder(),
+            ),
+          ),
+          const SizedBox(
+            height: tFormHeight - 20,
+          ),
+          TextFormField(
+            decoration: const InputDecoration(
+                prefixIcon: Icon(Icons.fingerprint),
+                labelText: tPassword,
+                hintText: tPassword,
+                border: OutlineInputBorder(),
+                suffixIcon: IconButton(
+                    onPressed: null, icon: Icon(Icons.remove_red_eye_sharp))),
+          ),
+          const SizedBox(
+            height: tFormHeight - 20,
+          ),
+          Align(
+              alignment: Alignment.centerRight,
+              child: TextButton(
+                  onPressed: () {}, child: const Text(tForgotPassword))),
+          SizedBox(
+            width: double.infinity,
+            child: ElevatedButton(
+                onPressed: () {}, child: Text(tLogin.toUpperCase())),
+          )
+        ],
+      ),
+    ));
+  }
+}
